+++ RolesScheduler/settings.py
@@ -138,8 +138,8 @@ USE_TZ = False
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/3.1/howto/static-files/
 
-STATIC_ROOT = os.path.join(BASE_DIR, 'staticfiles')
 STATIC_URL = '/static/'
+STATIC_ROOT = "/static/"
 
 # Redirect user to the home page after they login or logout
 LOGIN_REDIRECT_URL = 'agenda:index'