+++ agenda/models.py
@@ -22,6 +22,7 @@ class Meeting(models.Model):
     starttime = models.DateTimeField()
     wod = models.CharField(max_length=50, default=None, null=True)
     theme = models.CharField(max_length=50, default=None, null=True)
+    guestlist = models.CharField(max_length=256, default=None, null=True)
 
     def __str__(self):
         return f"{self.id}: {self.starttime}"
@@ -29,26 +30,26 @@ class Meeting(models.Model):
 
 class Rolelist(models.Model):
     meeting = models.ForeignKey("Meeting", on_delete=models.CASCADE, related_name="rolelist")
-    facilitator = models.ForeignKey("User", on_delete=models.CASCADE, related_name="facilitator", null=True)
-    zoom = models.ForeignKey("User", on_delete=models.CASCADE, related_name="zoom", null=True)
-    toastmaster = models.ForeignKey("User", on_delete=models.CASCADE, related_name="toastmaster", null=True)
-    geneval = models.ForeignKey("User", on_delete=models.CASCADE, related_name="geneval", null=True)
-    saa = models.ForeignKey("User", on_delete=models.CASCADE, related_name="saa", null=True)
-    chair = models.ForeignKey("User", on_delete=models.CASCADE, related_name="chair", null=True)
+    facilitator = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="facilitator", null=True)
+    zoom = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="zoom", null=True)
+    toastmaster = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="toastmaster", null=True)
+    geneval = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="geneval", null=True)
+    saa = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="saa", null=True)
+    chair = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="chair", null=True)
 
-    speaker1 = models.ForeignKey("User", on_delete=models.CASCADE, related_name="speaker1", null=True)
-    speaker2 = models.ForeignKey("User", on_delete=models.CASCADE, related_name="speaker2", null=True)
-    speaker3 = models.ForeignKey("User", on_delete=models.CASCADE, related_name="speaker3", null=True)
-    eval1 = models.ForeignKey("User", on_delete=models.CASCADE, related_name="eval1", null=True)
-    eval2 = models.ForeignKey("User", on_delete=models.CASCADE, related_name="eval2", null=True)
-    eval3 = models.ForeignKey("User", on_delete=models.CASCADE, related_name="eval3", null=True)
+    speaker1 = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="speaker1", null=True)
+    speaker2 = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="speaker2", null=True)
+    speaker3 = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="speaker3", null=True)
+    eval1 = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="eval1", null=True)
+    eval2 = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="eval2", null=True)
+    eval3 = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="eval3", null=True)
 
-    ttmaster = models.ForeignKey("User", on_delete=models.CASCADE, related_name="ttmaster", null=True)
-    tteval = models.ForeignKey("User", on_delete=models.CASCADE, related_name="tteval", null=True)
+    ttmaster = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="ttmaster", null=True)
+    tteval = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="tteval", null=True)
 
-    timer = models.ForeignKey("User", on_delete=models.CASCADE, related_name="timer", null=True)
-    ah_counter = models.ForeignKey("User", on_delete=models.CASCADE, related_name="ah_counter", null=True)
-    quizmaster = models.ForeignKey("User", on_delete=models.CASCADE, related_name="quizmaster", null=True)
+    timer = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="timer", null=True)
+    ah_counter = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="ah_counter", null=True)
+    quizmaster = models.ForeignKey("User", on_delete=models.SET_NULL, related_name="quizmaster", null=True)
 
     absences = models.BooleanField(default=0)